client/nextflow_tower_api_client/models/action_tower_action_event.py
import datetime
from typing import Any, Dict, List, Type, TypeVar, Union

from attrs import define as _attrs_define
from attrs import field as _attrs_field
from dateutil.parser import isoparse

from ..types import UNSET, Unset

T = TypeVar("T", bound="ActionTowerActionEvent")


@_attrs_define
class ActionTowerActionEvent:
    """
    Attributes:
        timestamp (Union[Unset, datetime.datetime]):
        workflow_id (Union[Unset, str]):
        discriminator (Union[Unset, str]):
    """

    timestamp: Union[Unset, datetime.datetime] = UNSET
    workflow_id: Union[Unset, str] = UNSET
    discriminator: Union[Unset, str] = UNSET
    additional_properties: Dict[str, Any] = _attrs_field(init=False, factory=dict)

    def to_dict(self) -> Dict[str, Any]:
        timestamp: Union[Unset, str] = UNSET
        if not isinstance(self.timestamp, Unset):
            timestamp = self.timestamp.isoformat()

        workflow_id = self.workflow_id
        discriminator = self.discriminator

        field_dict: Dict[str, Any] = {}
        field_dict.update(self.additional_properties)
        field_dict.update({})
        if timestamp is not UNSET:
            field_dict["timestamp"] = timestamp
        if workflow_id is not UNSET:
            field_dict["workflowId"] = workflow_id
        if discriminator is not UNSET:
            field_dict["discriminator"] = discriminator

        return field_dict

    @classmethod
    def from_dict(cls: Type[T], src_dict: Dict[str, Any]) -> T:
        d = src_dict.copy()
        _timestamp = d.pop("timestamp", UNSET)
        timestamp: Union[Unset, datetime.datetime]
        if isinstance(_timestamp, Unset):
            timestamp = UNSET
        else:
            timestamp = isoparse(_timestamp)

        workflow_id = d.pop("workflowId", UNSET)

        discriminator = d.pop("discriminator", UNSET)

        action_tower_action_event = cls(
            timestamp=timestamp,
            workflow_id=workflow_id,
            discriminator=discriminator,
        )

        action_tower_action_event.additional_properties = d
        return action_tower_action_event

    @property
    def additional_keys(self) -> List[str]:
        return list(self.additional_properties.keys())

    def __getitem__(self, key: str) -> Any:
        return self.additional_properties[key]

    def __setitem__(self, key: str, value: Any) -> None:
        self.additional_properties[key] = value

    def __delitem__(self, key: str) -> None:
        del self.additional_properties[key]

    def __contains__(self, key: str) -> bool:
        return key in self.additional_properties
